pyproject.toml
[tool.poetry]
name = "grupybnu-thumbs"
version = "0.1.0"
description = "Gerador de banners para os encontros do GruPy Blumenau"
authors = ["João Paulo Carvalho <email@jjpaulo2.dev.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
html2image = "^2.0.3"
pygithub = "^1.59.0"
typer = "^0.9.0"
questionary = "^1.10.0"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.4"
bandit = "^1.7.5"
mypy = "^1.4.1"

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.0"
pytest-randomly = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.main]
py-version = "3.8"

[tool.pylint."messages control"]
disable = [
    "missing-docstring"
]
